chore: fixed issue on s3 delete objects when there are no keys + refactored the websocket design to use a single service

+++ uasdm-server/src/main/java/gov/geoplatform/uasdm/remote/s3/S3RemoteFileService.java
@@ -445,14 +445,18 @@ public class S3RemoteFileService implements RemoteFileService
           }).collect(Collectors.toList());
 
       // Delete multiple objects in one request.
-      Delete del = Delete.builder().objects(keys).build();
 
-      DeleteObjectsRequest multiObjectDeleteRequest = DeleteObjectsRequest.builder() //
-          .bucket(bucketName) //
-          .delete(del) //
-          .build();
+      if (keys.size() > 0)
+      {
+        Delete del = Delete.builder().objects(keys).build();
 
-      client.deleteObjects(multiObjectDeleteRequest);
+        DeleteObjectsRequest multiObjectDeleteRequest = DeleteObjectsRequest.builder() //
+            .bucket(bucketName) //
+            .delete(del) //
+            .build();
+
+        client.deleteObjects(multiObjectDeleteRequest);
+      }
     }
     catch (SdkClientException e)
     {
